changed chart data method to return an int

+++ TaskManagerRemake.Domain/Services/PerformanceTab/IPerformanceItem.cs
@@ -11,7 +11,7 @@ namespace TaskManagerRemake.Domain.Services.PerformanceTab
     {
         string GetTabTitle();
         string GetTabSpecs();
-        double GetDataForChart();
+        int GetDataForChart();
         List<PerformanceStat> GetDynamicStats();
         List<PerformanceStat> GetStaticStats();
     }
